API/requirements.txt
Flask==1.1.2
Flask-CORS==3.0.10
Pillow==8.1.0
numpy==1.19.5